+++ grobid-core/src/main/java/org/grobid/core/engines/tagging/TaggerFactory.java
@@ -20,34 +20,22 @@ import org.slf4j.LoggerFactory;
 public class TaggerFactory {
     public static final Logger LOGGER = LoggerFactory.getLogger(TaggerFactory.class);
 
-    private static Map<GrobidModel, GenericTagger> cache = new HashMap<GrobidModel, GenericTagger>();
+    private static Map<GrobidModel, GenericTagger> cache = new HashMap<>();
+
+    private TaggerFactory() {}
 
     public static synchronized GenericTagger getTagger(GrobidModel model) {
-        GenericTagger t = cache.get(model);
-        if (t == null) {
-            switch (GrobidProperties.getGrobidCRFEngine(model)) {
-                case CRFPP:
-                    t = new CRFPPTagger(model);
-                    break;
-                case WAPITI:
-                    t = new WapitiTagger(model);
-                    break;
-                case DELFT:
-                    t = new DeLFTTagger(model);
-                    break;
-                default:
-                    throw new IllegalStateException("Unsupported Grobid sequence labelling engine: " + GrobidProperties.getGrobidCRFEngine());
-            }
-            cache.put(model, t);
-        }
-        return t;
+        return getTagger(model, GrobidProperties.getGrobidCRFEngine());
     }
 
     public static synchronized GenericTagger getTagger(GrobidModel model, GrobidCRFEngine engine) {
         GenericTagger t = cache.get(model);
         if (t == null) {
-            if (engine != null) {
-                System.out.println(engine);
+            if(model.equals(GrobidModels.DUMMY)) {
+                return new DummyTagger(model);
+            }
+
+            if(engine != null) {
                 switch (engine) {
                     case CRFPP:
                         t = new CRFPPTagger(model);
@@ -55,27 +43,24 @@ public class TaggerFactory {
                     case WAPITI:
                         t = new WapitiTagger(model);
                         break;
-                    case DELFT:                        
+                    case DELFT:
                         // be sure the native JEP lib can be loaded
-                        try {
-                            String libraryFolder = LibraryLoader.getLibraryFolder();
-                            System.out.println(libraryFolder);
-                            LibraryLoader.addLibraryPath(libraryFolder);
-                        } catch (Exception e) {
-                            LOGGER.info("Loading JEP native library for DeLFT failed", e);
-                        }
+//                        try {
+//                            String libraryFolder = LibraryLoader.getLibraryFolder();
+//                            System.out.println(libraryFolder);
+//                            LibraryLoader.addLibraryPath(libraryFolder);
+//                        } catch (Exception e) {
+//                            LOGGER.info("Loading JEP native library for DeLFT failed", e);
+//                        }
 
-                        // if model is fulltext or segmentation we use currently WAPITI as fallback because they
-                        // are not covered by DeLFT for the moment
-                        if (model.getModelName().equals("fulltext") || model.getModelName().equals("segmentation"))
-                            t = new WapitiTagger(model);
-                        else    
-                            t = new DeLFTTagger(model);
+                        t = new DeLFTTagger(model);
                         break;
                     default:
                         throw new IllegalStateException("Unsupported Grobid sequence labelling engine: " + GrobidProperties.getGrobidCRFEngine());
                 }
                 cache.put(model, t);
+            } else {
+                throw new IllegalStateException("Unsupported or null Grobid sequence labelling engine: " + GrobidProperties.getGrobidCRFEngine());
             }
         }
         return t;