+++ grobid-core/src/main/java/org/grobid/core/utilities/GrobidProperties.java
@@ -16,6 +16,7 @@ import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Enumeration;
+import java.util.Map;
 import java.util.Properties;
 
 /**
@@ -303,6 +304,8 @@ public class GrobidProperties {
             throw new GrobidPropertyException("Cannot open file of grobid properties" + getGrobidPropertiesPath().getAbsolutePath(), exp);
         }
 
+        getProps().putAll(getEnvironmentVariableOverrides(System.getenv()));
+
         initializePaths();
         //checkProperties();
         loadPdf2XMLPath();
@@ -334,6 +337,14 @@ public class GrobidProperties {
             }
         }
         return GROBID_VERSION;
+    }
+
+    protected static Map<String, String> getEnvironmentVariableOverrides(Map<String, String> environmentVariablesMap) {
+        Map<String, String> properties = new EnvironmentVariableProperties(
+            environmentVariablesMap, "GROBID__"
+        ).getProperties();
+        LOGGER.info("environment variables overrides: {}", properties);
+        return properties;
     }
 
     /**